Merge 2025.3-dev into master for release

See merge request kalilinux/nethunter/apps/kali-nethunter-term!35

+++ nhterm/src/main/java/com/offsec/nhterm/setup/SetupThread.java
@@ -58,7 +58,7 @@ final class SetupThread extends Thread {
         int totalBytes = sourceConnection.getSize();
 
         while ((zipEntry = zipInput.getNextEntry()) != null) {
-          totalReadBytes += zipEntry.getCompressedSize();
+          totalReadBytes += (int) zipEntry.getCompressedSize();
 
           final int totalReadBytesFinal = totalReadBytes;
           final int totalBytesFinal = totalBytes;
